Code modified to use a shared viewmodel to transfer data between fragments
