+++ app/src/main/java/com/dheeraj/news/presentation/view/NewsActivity.kt
@@ -1,15 +1,22 @@
 package com.dheeraj.news.presentation.view
 
 import android.os.Bundle
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.viewModels
 import com.dheeraj.news.R
+import com.dheeraj.news.presentation.viewmodel.NewsViewModel
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class NewsActivity : AppCompatActivity() {
 
+    private val newsViewModel: NewsViewModel by viewModels()
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_news)
     }
+
+    fun getViewModelInstance() = newsViewModel
 }
