+++ app/src/main/java/com/dheeraj/news/presentation/view/NewsArticleFragment.kt
@@ -8,24 +8,21 @@ import android.widget.ImageView
 import android.widget.Toast
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
+import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Observer
-import androidx.navigation.fragment.navArgs
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 import com.dheeraj.news.R
+import com.dheeraj.news.data.util.Resource
 import com.dheeraj.news.databinding.FragmentNewsArticleBinding
 import com.dheeraj.news.domain.model.NewsArticle
 import com.dheeraj.news.presentation.viewmodel.NewsViewModel
-import com.dheeraj.news.data.util.Resource
-import com.dheeraj.news.presentation.viewmodel.NewsArticleViewModel
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class NewsArticleFragment : Fragment() {
 
-    private val args: NewsArticleFragmentArgs by navArgs()
-    private val newsArticleViewModel: NewsArticleViewModel by viewModels()
+    private val newsArticleViewModel: NewsViewModel by activityViewModels()
     private lateinit var newsArticleBinding: FragmentNewsArticleBinding
 
     override fun onCreateView(
@@ -35,7 +32,6 @@ class NewsArticleFragment : Fragment() {
     ): View? {
         newsArticleBinding =
             DataBindingUtil.inflate(inflater, R.layout.fragment_news_article, container, false)
-        newsArticleViewModel.getLikesAndComments(args.newsArticle)
         return newsArticleBinding.root
     }
