+++ app/src/main/java/com/dheeraj/news/presentation/view/NewsFragment.kt
@@ -7,6 +7,7 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import androidx.navigation.fragment.findNavController
@@ -22,7 +23,7 @@ import dagger.hilt.android.AndroidEntryPoint
 @AndroidEntryPoint
 class NewsFragment : Fragment(), NewsArticleListAdapter.Interaction {
 
-    private val newsViewModel: NewsViewModel by viewModels()
+    private val newsViewModel: NewsViewModel by activityViewModels()
     private lateinit var fragmentNewsBinding: FragmentNewsBinding
     private lateinit var newsArticleListAdapter: NewsArticleListAdapter
 
@@ -81,7 +82,7 @@ class NewsFragment : Fragment(), NewsArticleListAdapter.Interaction {
     }
 
     override fun onItemSelected(position: Int, item: NewsArticle) {
-        val action = NewsFragmentDirections.actionLoadNewsArticle(item)
-        findNavController().navigate(action)
+        newsViewModel.getLikesAndComments(item)
+        findNavController().navigate(R.id.action_loadNewsArticle)
     }
 }
