+++ app/src/test/java/com/dheeraj/news/viewmodel/NewsViewModelTest.kt
@@ -2,6 +2,8 @@ package com.dheeraj.news.viewmodel
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.dheeraj.news.data.repository.FakeNewsRepository
+import com.dheeraj.news.domain.model.NewsArticle
+import com.dheeraj.news.domain.usecase.GetLikesAndCommentsUseCase
 import com.dheeraj.news.domain.usecase.GetNewsTopHeadlinesUseCase
 import com.dheeraj.news.presentation.viewmodel.NewsViewModel
 import com.dheeraj.news.util.TestCoroutineRule
@@ -23,6 +25,8 @@ class NewsViewModelTest {
 
     private lateinit var getNewsTopHeadlinesUseCase: GetNewsTopHeadlinesUseCase
 
+    private lateinit var getLikesAndCommentsUseCase: GetLikesAndCommentsUseCase
+
     private lateinit var fakeNewsRepository: FakeNewsRepository
 
     private lateinit var newsViewModel: NewsViewModel
@@ -31,8 +35,10 @@ class NewsViewModelTest {
     fun setUp() {
         fakeNewsRepository = FakeNewsRepository()
         getNewsTopHeadlinesUseCase = GetNewsTopHeadlinesUseCase(fakeNewsRepository)
+        getLikesAndCommentsUseCase = GetLikesAndCommentsUseCase(fakeNewsRepository)
         newsViewModel = NewsViewModel(
             getNewsTopHeadlinesUseCase,
+            getLikesAndCommentsUseCase,
             testCoroutineRule.testCorutineDispatcher
         )
     }
@@ -53,6 +59,26 @@ class NewsViewModelTest {
             newsViewModel.getNewsTopHeadlines()
             val newsArticles = newsViewModel.newsArticlesLiveData.getOrAwaitValue()
             assert(newsArticles.data == null)
+        }
+    }
+
+    @Test
+    fun getCommentsAndLikes_expectedResult() {
+        testCoroutineRule.runBlockingTest {
+            newsViewModel.getLikesAndComments(NewsArticle(articleId = "https://www.cbsnews.com/live-updates/boulder-shooting-colorado-2021-03-23/\""))
+            val newsArticle = newsViewModel.newsArticleLiveData.getOrAwaitValue()
+            assert(newsArticle.data?.likes == 56)
+            assert(newsArticle.data?.comments == 56)
+        }
+    }
+
+    @Test
+    fun getCommentsAndLikes_errorResult() {
+        fakeNewsRepository.setReturnError(true)
+        testCoroutineRule.runBlockingTest {
+            newsViewModel.getLikesAndComments(NewsArticle())
+            val newsArticle = newsViewModel.newsArticleLiveData.getOrAwaitValue()
+            assert(newsArticle.data?.likes == null && newsArticle.data?.comments == null)
         }
     }
