+++ build.gradle
@@ -8,7 +8,6 @@ buildscript {
     classpath "com.android.tools.build:gradle:4.1.2"
     classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:${Versions.kotlin_version}"
     classpath "com.google.dagger:hilt-android-gradle-plugin:${Versions.hilt}"
-    classpath "androidx.navigation:navigation-safe-args-gradle-plugin:${Versions.navigation_fragment}"
 
     // NOTE: Do not place your application dependencies here; they belong
     // in the individual module build.gradle files
